Improve prompt building
This fixes some spacing bugs and problems with segments not hiding, and follows the building pattern used by the original theme (post overhaul). There should be some speed improvements as well. However, I kept the git tracking the same because this method is faster. Hg may or may not work.